Delete duplicated network
