+++ networks/nova/Make.package
@@ -5,6 +5,7 @@ ifeq ($(USE_REACT),TRUE)
   CEXE_headers += actual_network.H
   CEXE_headers += tfactors.H
   CEXE_headers += partition_functions.H
+  CEXE_sources += partition_functions_data.cpp
   CEXE_headers += actual_rhs.H
   CEXE_headers += reaclib_rates.H
   CEXE_headers += table_rates.H
