+++ networks/nova/inputs.burn_cell.VODE
@@ -37,3 +37,7 @@ unit_test.X10 = 0.0
 unit_test.X11 = 0.0
 unit_test.X12 = 0.0
 unit_test.X13 = 0.0
+unit_test.X14 = 0.0
+unit_test.X15 = 0.0
+unit_test.X16 = 0.0
+unit_test.X17 = 0.0
