+++ networks/nova/partition_functions_data.cpp
@@ -1,0 +1,14 @@
+#include <AMReX_Array.H>
+#include <string>
+#include <table_rates.H>
+#include <AMReX_Print.H>
+
+#include <partition_functions.H>
+
+using namespace amrex;
+
+namespace part_fun {
+
+
+}
+
