+++ networks/nova/pynucastro.net
@@ -1,5 +1,9 @@
 hydrogen-1                H1        1.0    1.0
+hydrogen-2                H2        2.0    1.0
+helium-3                  He3       3.0    2.0
 helium-4                  He4       4.0    2.0
+beryllium-7               Be7       7.0    4.0
+boron-8                   B8        8.0    5.0
 carbon-12                 C12      12.0    6.0
 carbon-13                 C13      13.0    6.0
 nitrogen-13               N13      13.0    7.0
