--- networks/nova2/Make.package
@@ -1,14 +1,0 @@
-CEXE_headers += network_properties.H
-
-ifeq ($(USE_REACT),TRUE)
-  CEXE_sources += actual_network_data.cpp
-  CEXE_headers += actual_network.H
-  CEXE_headers += tfactors.H
-  CEXE_headers += partition_functions.H
-  CEXE_headers += actual_rhs.H
-  CEXE_headers += reaclib_rates.H
-  CEXE_headers += table_rates.H
-  CEXE_sources += table_rates_data.cpp
-  USE_SCREENING = TRUE
-  USE_NEUTRINOS = TRUE
-endif
