--- networks/nova2/_parameters
@@ -1,2 +1,0 @@
-@namespace: network
-
