--- networks/nova2/actual_network_data.cpp
@@ -1,46 +1,0 @@
-#include <actual_network.H>
-
-
-#ifdef NSE_NET
-namespace NSE_INDEX
-{
-    AMREX_GPU_MANAGED amrex::Array2D<int, 1, Rates::NumRates, 1, 7, Order::C> rate_indices {
-        -1, -1, 8, -1, -1, 7, -1,
-        -1, -1, 11, -1, -1, 9, -1,
-        -1, -1, 12, -1, -1, 10, -1,
-        -1, -1, 15, -1, -1, 14, -1,
-        -1, -1, 5, -1, 3, 3, -1,
-        -1, 0, 0, -1, -1, 1, -1,
-        -1, 0, 0, -1, -1, 1, -1,
-        -1, 0, 1, -1, -1, 2, -1,
-        -1, 1, 1, -1, -1, 3, -1,
-        -1, 0, 2, -1, -1, 3, -1,
-        -1, 2, 3, -1, -1, 4, -1,
-        -1, 0, 4, -1, -1, 5, -1,
-        -1, 0, 6, -1, -1, 8, -1,
-        -1, 3, 6, -1, -1, 13, -1,
-        -1, 0, 7, -1, -1, 9, -1,
-        -1, 0, 8, -1, -1, 11, -1,
-        -1, 0, 9, -1, -1, 12, -1,
-        -1, 3, 9, -1, -1, 16, -1,
-        -1, 0, 10, -1, -1, 13, -1,
-        -1, 0, 13, -1, -1, 15, -1,
-        -1, 0, 14, -1, -1, 16, -1,
-        -1, 1, 2, -1, 0, 3, -1,
-        -1, 3, 8, -1, 0, 13, -1,
-        -1, 0, 10, -1, 3, 6, -1,
-        -1, 3, 11, -1, 0, 15, -1,
-        -1, 0, 14, -1, 3, 9, -1,
-        -1, 0, 16, -1, 3, 12, -1,
-        -1, 2, 2, 0, 0, 3, -1,
-        -1, 1, 4, 0, 3, 3, -1,
-        -1, 2, 4, 0, 0, 3, -1,
-        3, 3, 3, -1, -1, 6, -1
-    };
-}
-#endif
-
-void actual_network_init()
-{
-
-}
