--- networks/nova2/inputs.burn_cell.VODE
@@ -1,43 +1,0 @@
-unit_test.run_prefix = "react_pynucastro_"
-
-unit_test.small_temp = 1e5
-unit_test.small_dens = 1e5
-
-integrator.burner_verbose = 0
-
-# Set which jacobian to use
-# 1 = analytic jacobian
-# 2 = numerical jacobian
-integrator.jacobian = 1
-
-integrator.renormalize_abundances = 0
-
-integrator.rtol_spec = 1.0e-6
-integrator.rtol_enuc = 1.0e-6
-integrator.atol_spec = 1.0e-6
-integrator.atol_enuc = 1.0e-6
-
-
-unit_test.tmax  = 1.0
-unit_test.nsteps = 1000
-
-unit_test.density       = 1.0e7
-unit_test.temperature   = 1.0e8
-
-unit_test.X1 = 1.0
-unit_test.X2 = 0.0
-unit_test.X3 = 0.0
-unit_test.X4 = 0.0
-unit_test.X5 = 0.0
-unit_test.X6 = 0.0
-unit_test.X7 = 0.0
-unit_test.X8 = 0.0
-unit_test.X9 = 0.0
-unit_test.X10 = 0.0
-unit_test.X11 = 0.0
-unit_test.X12 = 0.0
-unit_test.X13 = 0.0
-unit_test.X14 = 0.0
-unit_test.X15 = 0.0
-unit_test.X16 = 0.0
-unit_test.X17 = 0.0
