--- networks/nova2/partition_functions.H
@@ -1,167 +1,0 @@
-#ifndef PARTITION_FUNCTIONS_H
-#define PARTITION_FUNCTIONS_H
-
-#include <AMReX_REAL.H>
-#include <AMReX_Array.H>
-
-#include <tfactors.H>
-#include <fundamental_constants.H>
-#include <network_properties.H>
-
-using namespace amrex;
-using namespace Species;
-
-namespace part_fun {
-
-
-
-    // interpolation routine
-
-    template <int npts>
-    AMREX_GPU_HOST_DEVICE AMREX_INLINE
-    void interpolate_pf(const amrex::Real t9, const amrex::Real (&temp_array)[npts], const amrex::Real (&pf_array)[npts],
-                        amrex::Real& pf, amrex::Real& dpf_dT) {
-
-        if (t9 >= temp_array[0] && t9 < temp_array[npts-1]) {
-
-            // find the largest temperature element <= t9 using a binary search
-
-            int left = 0;
-            int right = npts;
-
-            while (left < right) {
-                int mid = (left + right) / 2;
-                if (temp_array[mid] > t9) {
-                    right = mid;
-                } else {
-                    left = mid + 1;
-                }
-            }
-
-            const int idx = right - 1;
-
-            // now we have temp_array[idx] <= t9 < temp_array[idx+1]
-
-            // construct the slope -- this is (log10(pf_{i+1}) - log10(pf_i)) / (T_{i+1} - T_i)
-
-            amrex::Real slope = (pf_array[idx+1] - pf_array[idx]) / (temp_array[idx+1] - temp_array[idx]);
-
-            // find the PF
-
-            amrex::Real log10_pf = pf_array[idx] + slope * (t9 - temp_array[idx]);
-            pf = std::pow(10.0_rt, log10_pf);
-
-            // find the derivative (with respect to T, not T9)
-
-            amrex::Real dpf_dT9 = pf * M_LN10 * slope;
-            dpf_dT = dpf_dT9 / 1.e9_rt;
-
-        } else {
-
-            // T < the smallest T or >= the largest T in the partition function table
-            pf = 1.0;
-            dpf_dT = 0.0;
-
-        }
-
-    }
-
-    struct pf_cache_t {
-        // Store the coefficient and derivative adjacent in memory, as they're
-        // always accessed at the same time.
-        // The entries will be default-initialized to zero, which is fine since
-        // log10(x) is never zero.
-        amrex::Array2D<amrex::Real, 1, NumSpecTotal, 1, 2, Order::C> data{};
-    };
-
-}
-
-// main interface
-
-AMREX_GPU_HOST_DEVICE AMREX_INLINE
-void get_partition_function(const int inuc, [[maybe_unused]] const tf_t& tfactors,
-                            amrex::Real& pf, amrex::Real& dpf_dT) {
-
-    // inuc is the 1-based index for the species
-
-    switch (inuc) {
-
-
-    default:
-
-        pf = 1.0_rt;
-        dpf_dT = 0.0_rt;
-
-    }
-
-}
-
-AMREX_GPU_HOST_DEVICE AMREX_INLINE
-void get_partition_function_cached(const int inuc, const tf_t& tfactors,
-                                   part_fun::pf_cache_t& pf_cache,
-                                   amrex::Real& pf, amrex::Real& dpf_dT) {
-    if (pf_cache.data(inuc, 1) != 0.0_rt) {
-        // present in cache
-        amrex::ignore_unused(tfactors);
-        pf = pf_cache.data(inuc, 1);
-        dpf_dT = pf_cache.data(inuc, 2);
-    } else {
-        get_partition_function(inuc, tfactors, pf, dpf_dT);
-        pf_cache.data(inuc, 1) = pf;
-        pf_cache.data(inuc, 2) = dpf_dT;
-    }
-}
-
-// spins
-
-AMREX_GPU_HOST_DEVICE AMREX_INLINE
-constexpr amrex::Real get_spin_state(const int inuc) {
-
-    amrex::Real spin = -1.0;
-
-    switch (inuc) {  // NOLINT(bugprone-switch-missing-default-case)
-
-    case He4:
-    case C12:
-    case O14:
-    case O16:
-        spin = 1;
-        break;
-
-    case H1:
-    case He3:
-    case C13:
-    case N13:
-    case N15:
-    case O15:
-        spin = 2;
-        break;
-
-    case H2:
-    case N14:
-    case F18:
-        spin = 3;
-        break;
-
-    case Be7:
-        spin = 4;
-        break;
-
-    case B8:
-        spin = 5;
-        break;
-
-    case O17:
-    case F17:
-        spin = 6;
-        break;
-
-
-    }
-
-    return spin;
-
-}
-
-
-#endif
