--- networks/nova2/pynucastro.net
@@ -1,17 +1,0 @@
-hydrogen-1                H1        1.0    1.0
-hydrogen-2                H2        2.0    1.0
-helium-3                  He3       3.0    2.0
-helium-4                  He4       4.0    2.0
-beryllium-7               Be7       7.0    4.0
-boron-8                   B8        8.0    5.0
-carbon-12                 C12      12.0    6.0
-carbon-13                 C13      13.0    6.0
-nitrogen-13               N13      13.0    7.0
-nitrogen-14               N14      14.0    7.0
-nitrogen-15               N15      15.0    7.0
-oxygen-14                 O14      14.0    8.0
-oxygen-15                 O15      15.0    8.0
-oxygen-16                 O16      16.0    8.0
-oxygen-17                 O17      17.0    8.0
-fluorine-17               F17      17.0    9.0
-fluorine-18               F18      18.0    9.0
