--- networks/nova2/table_rates_data.cpp
@@ -1,23 +1,0 @@
-#include <AMReX_Array.H>
-#include <string>
-#include <table_rates.H>
-#include <AMReX_Print.H>
-
-using namespace amrex;
-
-namespace rate_tables
-{
-
-
-}
-
-
-void init_tabular()
-{
-
-    amrex::Print() << "reading in network electron-capture / beta-decay tables..." << std::endl;
-
-    using namespace rate_tables;
-
-
-}
